Move test to test document
